Move block list merging from master to worker
Currently on worker register and heartbeat, the worker will flatten a
`Map<BlockStoreLocation, List<Long>>` to a
`List<LocationBlockIdListEntry>` and send to the master.
`LocationBlockIdListEntry` is a key-value pair
`BlockStoreLocationProto:BlockIdList`.

The problem with this approach is `BlockStoreLocation` is a tuple
`<tier,medium,dirIndex>` whereas `BlockStoreLocationProto` is a pair
`<tier,medium>`. Basically in map flattening, the
`List<LocationBlockIdListEntry>` contains more than one
`LocationBlockIdListEntry` with the same
`BlockStoreLocationProto:BlockIdList`. When the master receives the
list, it will reconstruct the map using `BlockStoreLocationProto` as the
key. At this step, there will be key collision because the worker can
have more directories with the same `<tier, medium>`. On collision, the
master will merge the block lists, which can induce huge list
reconstructions.

This change ensures the block lists of the same `<tier,medium>` (from
different directories) are merged before sending to the master.

pr-link: Alluxio/alluxio#13390
change-id: cid-3dcc4b0f0ac37a44d5460fe12de09218261f0522