Refactor MasterWorkerInfo
Currently all the worker metadata are wrapped in `MasterWorkerInfo`
object on the master side. This includes worker metadata, usage and the
various block lists. The existing `DefaultBlockMaster` does a
`synchronized (MasterWorkerInfo)` locking on this object on every read
or write operation on any field in this object. This locking is
excessive.

By refactoring this `MasterWorkerInfo` object and refactoring the
locking mechanism, the `DefaultBlockMaster` concurrency will be
improved, as multiple operations will be able to happen on the same
worker.

That said, this PR aims at the following:
1. Refactor `MasterWorkerInfo` object. Break this into multiple parts
which can be individually locked and updated.
2. Refactor the locking to lock individual parts on the object instead
of the whole object. RW Locking should also be applied.

pr-link: Alluxio/alluxio#13384
change-id: cid-5cd3ab31ee5c9fc24433b35485f2382e22a9a09f