Fix the cosn shaded pom
Fix some errors in the pom.xml of the cosn shaded module, which causes
the cosn shaded jar not to be packaged correctly.

Before the repair, the dependent components that need to be included
were not correctly specified in the pom file. Therefore, no code is
correctly packaged into the cosn's shaded jar:

![image](https://user-images.githubusercontent.com/4209464/119234304-cc5f0a80-bb5f-11eb-9671-c4275602f667.png)

After repairing, it can be packaged correctly:

![截屏2021-05-23 上午12 43
40](https://user-images.githubusercontent.com/4209464/119234339-0203f380-bb60-11eb-934c-9c841fdb6c2d.png)

pr-link: Alluxio/alluxio#13468
change-id: cid-b3fddb239d4601d0a6bd4c23b070442ab30c3555