Add journal checkpoint warnings on UI
### What changes are proposed in this pull request?

Adding warnings to the UI overview page when the journal has not
checkpointed in some time after reaching the journal entry checkpoint
period. Also, add the two checkpoint metrics (last checkpoint time and
entries since checkpoint) to the metrics tab.

New overview page warning:


![image](https://user-images.githubusercontent.com/3464508/123139176-c7190800-d40a-11eb-8deb-f6f444b35ad4.png)

Metrics page addition:


![image](https://user-images.githubusercontent.com/3464508/123139207-d13b0680-d40a-11eb-86dd-6c2a7d29d533.png)

### Why are the changes needed?

This is required to help alert users to potential issues with their
system if the journal has not checkpointed when it is supposed to.

### Does this PR introduce any user facing changes?

- Addition of a hidden property key to change the time period which
causes the warning to show up on the overview page. It is marked as
hidden because no one _should_ need to configure this value, but it may
be desirable. It is also useful for testing.
- Web UI additions (see screenshots). New section on the metrics page
and new potential warning in the overview.

pr-link: Alluxio/alluxio#13690
change-id: cid-d2b43991e86e92293a6c8124587bcc210382a5fa