Update and improve conformance of S3 API
### What changes are proposed in this pull request?

Improve conformance of the S3 API to the specification defined by AWS.

### Why are the changes needed?

The AWS CLI does not work for many standard operations (e.g. listing
buckets, setting max number of return keys, copying between buckets)

### Does this PR introduce any user facing changes?

This affects some of the values from the S3 API to be in-line with more
of the S3 REST API specification. It can potentially break backwards
compatibility with previous versions.

### PR Summary

This PR is a a serious of small changes and improvements to the S3
API which improves the conformance to the specification defined
by the AWS Rest API reference at
https://docs.aws.amazon.com/AmazonS3/latest/API/Type_API_Reference.html

A short summary of the changes:

- Implement the CopyObjectResult type for bucket-bucket transfers
- Make the ListAllMyBucketsResult conform to the spec. Added
creationDate and proper timestamps to results
- ListBucketOptions now includes the delimiter and encoding type
parameters
- the listBucketResult uses the CommonPrefix type and now includes
the delimiter and encoding type. Additionally, make sure keyCount
and truncation takes into account the common prefixes
- Update the S3RestServiceHandler to use Java 8 lambda language
features
- Update credentials to use the latest amazon authorization format
- A potentially breaking change which causes a ListAllMyBuckets result
to only return folders

pr-link: Alluxio/alluxio#13786
change-id: cid-297ff4411b461dd5331a5ed82c6f177df855c820