Fix the IndexOutOfBoundsException when reading one file
### What changes are proposed in this pull request?

This pr resolves the https://github.com/Alluxio/alluxio/issues/13824
that when the given buffer's length is larger than it read, it will
throw IndexOutOfBoundsException.

This occurs when the length of a given buffer minus offset is greater
than the actual read data.

### Why are the changes needed?
It's one bug

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#13849
change-id: cid-2d7d916699fada84c47e5cbbc38703c2016c4b9e