Support transfer leadership to a follower master in embedded journal
### What changes are proposed in this pull request?
Fixes #13499
Please outline the changes and how this PR fixes the issue.

### Why are the changes needed?
These changes are needed to give more flexibility in a Embedded Journal
HA cluster.

### Does this PR introduce any user facing changes?
The user will be able to use new CLI commands to designate which
secondary master will take over the quorum as the new leader.

pr-link: Alluxio/alluxio#13871
change-id: cid-f9f65e75a0563f4ca2b54d63e7f76bd52e20c17f