Fix extra preparations and cleanups in GetPinnedFileIdsBench
### What changes are proposed in this pull request?

Fix extra preparations and cleanups in GetPinnedFileIdsBench.

### Why are the changes needed?

`GetPinnedFileIdsBench` does not properly handle cases where the
benchmark is running in a local process or in a job worker, causing
unwanted extra runs of `prepare()` and `cleanup()`. This may cause the
expected number of test files to be incorrect and fail the benchmark.

### Does this PR introduce any user facing changes?

No.

pr-link: Alluxio/alluxio#13977
change-id: cid-d16c552718375ba200ceb3184e82289e39e6d822