Remove ramfs from size-checking condition
### What changes are proposed in this pull request?

Remove the check for the ramdisk size for the `ramfs` filesystem type on
the Alluxio worker.

### Why are the changes needed?

There has been a flaky bug recently where the worker will fail to start
because it measures the `ramfs` size as being smaller than Alluxio's
configured capacity. The source of this error likely comes from trying
to read the `size` option from the `mount` command output.

However, this check is actually useless and doesn't prevent us from
allocating more data than the mount point can handle. This is because
the `ramfs` type is actually unsized and can *always* consume up the
entire system's memory. It has no limits. From the kernel documentation
[[1]]:

> One downside of ramfs is you can keep writing data into it until you
fill
up all memory, and the VM can't free it because the VM thinks that files
should get written to backing store (rather than swap space), but ramfs
hasn't
got any backing store. Because of this, only root (or a trusted user)
should
be allowed write access to a ramfs mount.

> A ramfs derivative called tmpfs was created to add size limits, and
the ability
to write the data to swap space. Normal users can be allowed write
access to
tmpfs mounts. See Documentation/filesystems/tmpfs.txt for more
information.

Most modern kernels (v4+) don't even print the size option to the kernel
anyways when running `mount`. So this fix should help those running
kernels v3 and lower who may sometimes experience a failure to start the
worker when the size option is printed, but is essentially meaningless.
After this change we will still check the `tmpfs` size on startup, which
is size limited and should not be removed.

[1]:
https://www.kernel.org/doc/Documentation/filesystems/ramfs-rootfs-initramfs.txt

### Does this PR introduce any user facing changes?

No

pr-link: Alluxio/alluxio#13988
change-id: cid-8dbf2618a6d3f74d76de6b12ce9945b7719a3067