Register journal sequence number metric
### What changes are proposed in this pull request?

We need the sequence number information of log to monitor the cluster
status. The corresponding metrickey is provided in the source code, but
it is not registered to the metric system.

Please outline the changes and how this PR fixes the issue.

Register the metric key related to the sequence number to the metric
system

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#13956
change-id: cid-1a4c9ec039700e8fba6b4aa1a948b06f87c6fa84