Generate worker fuse helm yaml files
This PR is a continuation of
[https://github.com/Alluxio/alluxio/pull/13234](url)

To enable FUSE in worker container in k8s environment, the helm yaml
files need to be modified. According to
[https://github.com/Alluxio/alluxio/pull/13234](url), instead of
directly providing ready-to-use templates, we want to auto-generate
these templates with helm-generate.sh. User needs to include
"--enable-worker-fuse" in command line to turn worker fuse on.
The first commit is able to generate workerFuse container yaml files if
.Values.worker.fuseEnabled is true.
The second commit is able to get the parameter whether to enable
workerFuse from command line and pass it.

Fix Alluxio/new-contributor-tasks#587

pr-link: Alluxio/alluxio#13836
change-id: cid-1dba23f0abb94f5c87d73e5268a72a0dfdaa2fc8