Create smaller alpine and centos development docker image
### What changes are proposed in this pull request?
We introduce two dockerfiles: Dockerfile and Dockerfile-dev. Dockerfile
is for production use based on Alpine, where we try to minimize the size
of the image as much as possible. Dockerfile-dev is for development use
based on CentOS, where more tools are provided. dockerfile-common.sh
contains some common commands used both for Dockerfile and
Dockerfile-dev for better maintainability.

### Why are the changes needed?
On one hand, docker image with smaller size uses less time to build, is
easier to run, and takes less space, which is why we have the
Dockerfile. On the other hand, many users need to do further
development, where more tools are needed, which is why we have the
Dockerfile-dev.

### Does this PR introduce any user facing changes?
- To build the Dockerfile-dev, need to include `-t Dockerfile-dev`.
- Ubuntu, which is the base operating system for the old Dockerfile, is
replaced by Alpine in Dockerfile and CentOS in Dockerfile-dev

pr-link: Alluxio/alluxio#14035
change-id: cid-0462b212da6331aafd32b90ad2a554229c23267b