Update load command using new cache request
### What changes are proposed in this pull request?
This is a first version of draft PR. Add new cache request method to
blockWorker and use new method in loadCommand.
There's several TODO: add error message in cache request, add metrics
tracking cache status, etc...

### Why are the changes needed?
The load command currently uses the file streaming API to read data from
the worker, which also caches the data onto the worker. This is very
inefficient because the data has to be transferred to the client and
waste a lot of bandwidth. Using the cache API should be more lightweight
as it only instructs the worker to cache the block and not introduce any
data transfer to the client.
### Does this PR introduce any user facing changes?
probably add more option to Load command

pr-link: Alluxio/alluxio#13799
change-id: cid-f0839a70deef388ce77a94919e0add983093ceb8