Add property to skip listing broken symlinks on local UFS
### What changes are proposed in this pull request?

Add a property to skip broken symlinks when calling listStatus with the
LocalUnderFileSystem


### Why are the changes needed?

When a broken symlink appears when using a local UFS, all listings will
fail because of the additional call to `#readProperties` in
`LocalUndeFileSystem#listStatus`. This behavior can help alert users to
broken symlinks, but also breaks functionality until the symlink is
removed, which may be undesirable.

### Does this PR introduce any user facing changes?

A new option that can be used with a local UFS mount.

pr-link: Alluxio/alluxio#14044
change-id: cid-b091709b9db32c105bf69ef4681f35b208358782