Support running HA JobMaster process in intellij
### What changes are proposed in this pull request?
Add tools for intellij to start local HA JobMaster process., this pr is
a Is a supplement to pr #14042
### Why are the changes needed?
It is convenient to debug and test job master in intellij.
Please clarify why the changes are needed. For instance,
1. Add AlluxioJobMaster xml file to dev/intellij directory
2. Update the doc.
### Does this PR introduce any user facing changes?
No
Please list the user-facing changes introduced by your change, including
No

pr-link: Alluxio/alluxio#14120
change-id: cid-e55be3c958b813c2310de6e509d867a0018a4b21