Remove one constraint for listfile in FuseIOStressBench
In FuseIOStressBench, num-dirs is not needed for running ListFile
operation, but the logic compares number of threads and number of
directories, and throw error if the number of threads is larger than
number of directories. This constraint should not apply to ListFile
operation. This PR fixes this problem.

pr-link: Alluxio/alluxio#14202
change-id: cid-f4d553d4f2e39dcf6e156622f36db1f24588a681