Update alluxio-maven to use go1.12 + update cacerts
### What changes are proposed in this pull request and why are they
needed?

The alluxio-maven image is used to verify PR builds. Currently it is
installing go1.11 but the generate tarball go code uses
strings.ReplaceAll which is only available starting go1.12.

In addition, some commands that depend on curl/wget are failing due to
client certificates being out of date. This is fixed by upgrading the
ca-certificates package.

### Does this PR introduce any user facing changes?

They better not...

pr-link: Alluxio/alluxio#14244
change-id: cid-2abde50e7ae92481c0bae60f0085260a9c9f5343