The battle of Hadoop (Spark) vs zac is over
I have prevailed

Gosh this was an annoying bug. Turns out, the mapred
FileOutputCommitter (FOC) needs to determine if its
source or destination directories are files or directories.
It delegates this task to the S3A hadoop FS implementation
which is probably one of the most unintuitive pieces of
code I've ever had the pleasure of staring at for 2.5 hours.

Anyways, on with the battle; I finally figured out that the
s3a FS will determine if a file is a directory simply if the
path has a "/" at the end of it. Simple, right? WRONG. It only
uses this check if it previously stored the metadata of a
HeadObject request in the S3A FS's local cache. Otherwise
it delegates to an *actual* HeadObject request. However,
directories in S3 don't exist, so on a HeadObject to a
directory, S3A FS expects a 404 error, not a successful
response.

The S3RestServiceHandler was returning successful
responses on HeadObject for directories and files.
Returning a 404 on directories was enough to coax the
beast of the S3AFS impl to see the path as a directory, and
finally I was able to fix a minor rename bug -- which now
results in a successful output of the parquet files from
Spark. Amazing!