more headache with HeadObject
Had to do a lot of testing to see how the actual S3 API handles key
prefixes. This is undocumented in the AWS API, but a prefix (generally
representing a directory) will return 200 OK on the HeadObject request
but will have a 0 Content-Length header. The 0 Content-Length header
is crucial to indicating to the S3AFileSystem impl that the key does
in-fact represent a directory.