Fix listBucket bug
When listing objects we were not returning all of the status required.
This caused the FOC to not actually rename the files, and return an
empty destination directory.

I could not understand why we had the conditional in the first place
for the listStatus. I believe is was due to a recent change for the
ListObjectsV2 API support.