Add maxThroughput for job service
### What changes are proposed in this pull request?
Add MaxThroughput for job service, introduce several abstract classes in
stressBench so, later on, we can refactor other stressBench. Would
introduce another PR refactoring MaxThroughput(master) after
understanding the usage of that component.

### Why are the changes needed?
The reason we are doing refactoring is that there's lots of duplicate
code between job service maxThroughput and maxThroughput(master). And
there's also some duplication between the stress benchs.
Notice the key change we introduce is adding a class `GeneralParameter`.
We are introducing this class because we need some abstract Parameter
class with the operation member variable in MaxThroughput Summary. The
operation parameter needs a specific static converter to convert the
string to an Enum type. and we can not do resource injection or
template(since we have to put converter.class as a value to the
Parameter notation for `mOperation`). So we decide to go with an
alternative way to add an method to the Parameter class to get the
corresponding Operation.

### Does this PR introduce any user facing changes?
add maxThroughput command
`runClass alluxio.stress.cli.suite.JobServiceMaxThroughput --operation
NoOp --target-throughput 2000`

pr-link: Alluxio/alluxio#14207
change-id: cid-0b0d6bbbd02511951d3399c233ccea39b216934f