Save transfer of leadership from unchecked exceptions
..., transfers to current leader, and correctly report transfer errors
to users.

### What changes are proposed in this pull request?
Wrapping all of the `transferLeadership` code in a try block to
safeguard against unchecked exception. A bug is currently present where
some functions during the `RaftJournalSystem.transferLeadership` command
throw unchecked exceptions. Right now, the `allowed` bool does not get
reset to `true` and prevents all future transferLeadership attempts.

If the transferLeadership is initiated towards the current leader, then
the `allowed` bool is never reset to `true`, which prevents new
transfers.

Currently, if the `quorum elect` command receives an exception during a
transfer, it keeps retrying until the timeout. This behavior is
incorrect and the retry loop should end as soon as an exception is
encountered. This also correctly displays the error to the user instead
of a timeout exception for all types of errors.

### Does this PR introduce any user facing changes?
No.

pr-link: Alluxio/alluxio#14316
change-id: cid-54046c9406eb7d92e2307583dbd753bcd568827e