Fix distributedLoad after file is already loaded
### What changes are proposed in this pull request?
This PR introduces two changes.
1. Explicitly turn on passive caching when trying to load using a read
with localFirstPolicy, otherwise this read is an noop when the data is
already loaded in Alluxio. (It does a remote read)
2. Pass explicit worker information to the job worker to load the data,
this way, we can accurately load into the correct worker even when
locality is not clear. (Multiple workers per host or job worker and
worker are separate containers)

Fixes #14274

### Why are the changes needed?

When distributedLoad or setReplication runs when the file is already
loaded, nothing happens if passive caching is turned off.

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#14317
change-id: cid-61337566e09fa04deb5e0343f067a09341219113