Convert worker registration to a stream
### What changes are proposed in this pull request?

This change adds the following items:
1. Add an option for the worker register logic to be a streaming call
instead of unary
2. Add a stressbench implementation

As a side effect, the locking mechanism in `MasterWorkerInfo` is changed
from `ReentrantReadWriteLock` to `StampedLock`. See the discussion in
this PR for why.

### Why are the changes needed?

When the worker has a huge amount of blocks to report to the master, the
worker register call will be a huge unary RPC call. This incurs a lot of
heap allocation on the master side and it's mostly not GC-able before
the call finishes. Changing this to streaming makes it more GC-friendly.

Being able to do some GC during the stream can allow higher concurrency
too, so more workers can successfully register at the same time.

Streaming also enables more flow control logic that was impossible when
using a unary call. This is not included in this PR but may be added in
the future.

### Does this PR introduce any user facing changes?

The stream will hold the write locks on the `MasterWorkInfo` during the
stream. As a side effect, the Web UI and other readers of
`MasterWorkerInfo` will block when they are inquiring the status of all
workers. When the worker finishes registering, or when the stream times
out, the locks will be released and the readers will be unblocked. This
is one field to improve on in future changes.

pr-link: Alluxio/alluxio#14223
change-id: cid-815d1792d6a2380544f5f2ee11d9faf989bf31bc