Fix flakiness of ListAllMyBuckets test
### What changes are proposed in this pull request?

Previously, the test wasn't actually getting the real status and
comparing it to the result, so the test relied on millisecond-level
timing in order to run successfully. On constrained CI environments
this led to many failures.

Using the real status should alleviate the issue.

### Why are the changes needed?

alleviate flaky unit test

### Does this PR introduce any user facing changes?

no

pr-link: Alluxio/alluxio#14404
change-id: cid-2dac098b16057a59015267901b6a46d9c0109d4f