[DOCFIX] Update Flink documentation
In the english version of the flink doc, in the prerequisite section it
says `Flink has been set up and is running.` which sounds like there
should be a running cluster started by the Flink's
`bin/start-cluster.sh` command. However, after the changes are made to
the configuration of hadoop and flink, running the example in the doc
won't work unless the running flink cluster is stopped.

pr-link: Alluxio/alluxio#14411
change-id: cid-b647b7f454ef1c4baca566b167a759aa99e0b553