Add metric to monitor checkpoint
### What changes are proposed in this pull request?

Add metric to monitor checkpoint

### Why are the changes needed?

Refer to the implementation of
AlluxioMasterRestServiceHandler#getWebUIOverview. Although there will be
an alert on the web page when the checkpoint is abnormal, we want to use
a third-party operation and maintenance system to monitor this
situation, so we hope to expose this metric. When it is 1, the
monitoring system is triggered to send an alarm message .

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#14327
change-id: cid-d4ab48e20ab2a0069f178e1c4b55c611f6d2c887