Fix default value and description for inode cache property key
### What changes are proposed in this pull request?

The default value of `alluxio.master.metastore.inode.cache.max.size` is
computed at runtime based on the max memory of master JVM. However, the
doc for that prop entry is statically generated at build time, thus
reflecting the default value on the build machine.

### Why are the changes needed?

This PR fixes the description, so that the doc correctly shows the
formula used to compute the runtime default value, instead of a build
time constant.

### Does this PR introduce any user facing changes?

This is mostly doc fix.

pr-link: Alluxio/alluxio#14873
change-id: cid-086e63fb4281caaee6db6b4fa380f893e707d745