[DOCFIX] Add readme to csi driver
The CSI driver directory is lacking a README. We are trying to add
Alluxio CSI driver into kuberbetes csi drivers list documentation, and
the hyperlink to Alluxio CSI driver directory will be included. However
currently there is no explanation of anything in that directory. This PR
adds a README that points to our CSI documentation.

pr-link: Alluxio/alluxio#14937
change-id: cid-1c8f4a1ae2c6cce78f78787315e0399aa86c73ad