Fix flaky FileOutStreamIntegrationTest#cancelWrite
### What changes are proposed in this pull request?

Fix an integration test of FileOutStreamIntegrationTest#cancelWrite

### Why are the changes needed?

Original test setup intended to customize worker-master heartbeat
interval to speed up the test.
However, a bug was introduced in
https://github.com/Alluxio/alluxio/pull/10588/
which customized the wrong property key.
As a result, the grace period is only 0.5s as supposed to be 2s,
introducing flakiness due to lack of time
for block information to be propagated from worker to master.

### Does this PR introduce any user facing changes?

No

pr-link: Alluxio/alluxio#14958
change-id: cid-92393dfb9a8ac2c1de6b44cd111de788ad2e1b4b