Fix Flaky AlluxioMasterProcessTest.stopAfterStandbyTransition
### What changes are proposed in this pull request?

Fix Flaky
alluxio.master.AlluxioMasterProcessTest.stopAfterStandbyTransition

### Why are the changes needed?

The test itself is simple, it tries to simulate the process when the
master sets to a standby state.
And this test will check if the master cleans up as follows:
1. clean up RPC port.
2. clean up web port.
3. master switch to not running state.

However, it sleeps 10 seconds to wait for all cleanings to happen :(

In this pr, I just increase the sleep time to make this test less flaky
and we should avoid this test pattern in the future.

### Does this PR introduce any user facing changes?

no

pr-link: Alluxio/alluxio#14984
change-id: cid-4d4f88be07e3a9ad689ab7f6afdc2074829922d5