Fix how to attach to Alluxio process using remote debugger
Attaching a debugger to an Alluxio process was broken. This was due to
the fact that the `-agentlib` parameter was passed alongside the
`ALLUXIO_JAVA_OPTS`.
For example, when trying the attach to an Alluxio master process using
`suspend=y`, the `-agentlib` argument would be passed to
`alluxio.cli.GetConf` and would hang until a debugger is attached to it,
preventing the master from launching.
When setting `suspend=n` but assigning a port through the `address`
parameter, both `alluxio.cli.GetConf` and `alluxio.master.AlluxioMaster`
would both try to set up debug servers using the same port.
This PR decouples the `-agentlib` parameter into its own configurable
parameter to prevent this issue.

pr-link: Alluxio/alluxio#14991
change-id: cid-22c17e9933ee6f93fcc6293c6811a3fc3770c3cd