Remove usage of Helm hooks in Helm chart
### What changes are proposed in this pull request?

Removal of the Helm hook annotations in our Helm chart.

### Why are the changes needed?

The usage of Helm hooks for `ConfigMap` resources means that they cannot
be `helm upgrade`-ed. We _could_ add the `"helm.sh/hook": "pre-upgrade"`
annotation to address that, but I actually think these resources should
not be hooks in the first place.

From the [Helm hooks
documentation](https://helm.sh/docs/topics/charts_hooks/):
> **Hook resources are not managed with corresponding releases**
>
> The resources that a hook creates are currently not tracked or managed
as part of the release. Once Helm verifies that the hook has reached its
ready state, it will leave the hook resource alone. Garbage collection
of hook resources when the corresponding release is deleted may be added
to Helm 3 in the future, so any hook resources that must never be
deleted should be annotated with
[helm.sh/resource-policy:](http://helm.sh/resource-policy:) keep.

`ConfigMap` resources are definitely something that are coupled with
individual Helm releases, and it doesn't make sense to have to manage
their lifecycles independently of releases. Hooks would make sense for
one-time jobs or bootstrapping services, but not the currently affected
`ConfigMap`s.

### Does this PR introduce any user facing changes?

No user-facing changes.

pr-link: Alluxio/alluxio#15021
change-id: cid-bdc57891d324dfa62b682904b3097fea14045e19