Use increment-by-count instead of iteration for fail & cancel counters
### What changes are proposed in this pull request?
1. Use increment-by-counts instead of iteration for fail and cancel stat
counters
2. Refactor DistributedMetrics to skip for other non-supported job
types.

Please outline the changes and how this PR fixes the issue.
1. Previously there is a performance drop, which may be caused by the PR
https://github.com/Alluxio/alluxio/pull/14989. Inspecting the PR, code
changes for incrementing failure counters are added.
When the code runs for batch config, if the folder has lots of files and
the batch size is huge, incrementing failure counters in the iteration
will be slow.
Changing the iteration to adding the increment values directly will
avoid the long-running loop, and thus speed up the code to run.
2. Refactored the code to skip unsupported job types, so that the code
looks cleaner and would run without further going through the remaining
code body.
3. Added a unit test to test Complete status, removed some unnecessary
tests.

pr-link: Alluxio/alluxio#15044
change-id: cid-1d03317bce67f18ac446f75d6d8ec6d1890e2f0f