[SMALLFIX] Change the default base directory path of StressBench
### What changes are proposed in this pull request?

I changed the default baes directory of StressBench from
"alluxio://localhost:19998/" into "alluxio:///" .

### Why are the changes needed?

The former default base directory is localhost:19998, but we may not use
the port 19998 in some case. So the directory "alluxio:///" is more
reasonable.

### Does this PR introduce any user facing changes?

The default base directory is changed, but the user don't need care of
this since the former directory is included in the new directory

pr-link: Alluxio/alluxio#15082
change-id: cid-4cae85b94b5443747d95491ff567424e6fdce26d