[SMALLFIX] Add description to some StressBench tool
### What changes are proposed in this pull request?

Add description article to StressClientIOBench and StressWorkerBench to
introduce their function and usage. Add description to the fixed-count
parameter of StressMasterBench to describe its usage in some operation.

### Why are the changes needed?

The description will help the user find understand the StressBench tool
easier.

### Does this PR introduce any user facing changes?

There is no changes to the user.

pr-link: Alluxio/alluxio#15085
change-id: cid-587aa06c879650fe360e659011055216bb860d2b